Rename Android package
* Rename package to com.onesignal.cordova
