+++ src/android/com/onesignal/cordova/OneSignalController.java
@@ -1,4 +1,4 @@
-package com.plugin.gcm;
+package com.onesignal.cordova;
 
 import com.onesignal.OSDeviceState;
 import com.onesignal.OneSignal;
