+++ src/android/com/onesignal/cordova/OneSignalPush.java
@@ -25,7 +25,7 @@
  * THE SOFTWARE.
  */
 
-package com.plugin.gcm;
+package com.onesignal.cordova;
 
 import android.util.Log;
